Power_Service_Extesnsibility/PowerServiceDriver/src/requirements.txt
cloudshell-shell-core>=2.3.2,<2.4.0
cloudshell-automation-api>=7.1.0.3